Extend existing REST resource to support updating policy for existing applications CLM-23 CLM-143

+++ insight-brain-service/src/test/java/com/sonatype/insight/brain/service/AbstractBrainServiceTest.java
@@ -269,7 +269,14 @@ public abstract class AbstractBrainServiceTest
         for ( Application application : applicationsToDelete )
         {
             cleanupApplication( application );
-            applicationDAO.delete( application );
+            if (application.getId() != null)
+            {
+                application = applicationDAO.getById( application.getId() );
+                if (application != null)
+                {
+                    applicationDAO.delete( application );
+                }
+            }
         }
         applicationsToDelete.clear();
     }
